Cleanup settings menu slightly
Rename extra to Additional
Move CFW Settings into Additional
Move the non settings to the bottom of additional (they dont belong here but where to put them?)
