+++ App/PyUI/main-ui/menus/settings/extra_settings_menu.py
@@ -8,6 +8,7 @@ from display.on_screen_keyboard import OnScreenKeyboard
 from games.utils.box_art_resizer import BoxArtResizer
 from menus.language.language import Language
 from menus.settings import settings_menu
+from menus.settings.cfw_system_settings_menu_for_category import CfwSystemSettingsMenuForCategory
 from menus.settings.controller_settings_menu import ControllerSettingsMenu
 from menus.settings.display_settings_menu import DisplaySettingsMenu
 from menus.settings.game_art_display_settings_menu import GameArtDisplaySettingsMenu
@@ -19,6 +20,8 @@ from menus.settings.modes_menu import ModesMenu
 from menus.settings.time_settings_menu import TimeSettingsMenu
 from themes.theme import Theme
 from utils.boxart.box_art_scraper import BoxArtScraper
+from utils.cfw_system_config import CfwSystemConfig
+from utils.logger import PyUiLogger
 from utils.py_ui_config import PyUiConfig
 from views.grid_or_list_entry import GridOrListEntry
 
@@ -84,6 +87,13 @@ class ExtraSettingsMenu(settings_menu.SettingsMenu):
         if (ControllerInput.A == input):
             BoxArtResizer.patch_boxart()
             
+
+    def launch_settings_for_category(self,input, category):
+        if(ControllerInput.A == input):
+            CfwSystemSettingsMenuForCategory(category).show_menu()
+
+
+
     def build_options_list(self):
         option_list = []
         
@@ -135,19 +145,6 @@ class ExtraSettingsMenu(settings_menu.SettingsMenu):
                     )
             )
         
-        if(Device.supports_image_resizing()):
-            option_list.append(
-                GridOrListEntry(
-                    primary_text="Optimize Boxart",
-                    value_text=None,
-                    image_path=None,
-                    image_path_selected=None,
-                    description=None,
-                    icon=None,
-                    value=self.resize_boxart
-                )
-            )        
-
         if(PyUiConfig.allow_pyui_game_switcher()):
             option_list.append(
                     GridOrListEntry(
@@ -200,7 +197,7 @@ class ExtraSettingsMenu(settings_menu.SettingsMenu):
 
         option_list.append(
             GridOrListEntry(
-                primary_text="Language",
+                primary_text="Language Settings",
                 value_text=None,
                 image_path=None,
                 image_path_selected=None,
@@ -223,6 +220,47 @@ class ExtraSettingsMenu(settings_menu.SettingsMenu):
                             value=self.launch_stock_os_menu
                         )
                 )
+
+        
+        for category in CfwSystemConfig.get_categories():
+            menu_options = CfwSystemConfig.get_menu_options(category=category)
+            contains_entry_for_device = False
+            for name, option in menu_options.items():
+                PyUiLogger.get_logger().info(f"{option}")
+                devices = option.get('devices')
+                supported_device = not devices or Device.get_device_name() in devices
+                if(supported_device):
+                    contains_entry_for_device = True
+                    break
+
+            if(contains_entry_for_device):
+                option_list.append(
+                        GridOrListEntry(
+                                primary_text=category,
+                                value_text=None,
+                                image_path=None,
+                                image_path_selected=None,
+                                description=None,
+                                icon=None,
+                                value=lambda 
+                                    input_value, 
+                                    category=category: self.launch_settings_for_category(input_value, category)
+                            )
+                    )
+
+
+        if(Device.supports_image_resizing()):
+            option_list.append(
+                GridOrListEntry(
+                    primary_text="Optimize Boxart",
+                    value_text=None,
+                    image_path=None,
+                    image_path_selected=None,
+                    description=None,
+                    icon=None,
+                    value=self.resize_boxart
+                )
+            )        
 
         option_list.append(
             GridOrListEntry(
